+++ src/app/project/ontology-classes/ontology-class-instance/ontology-class-instance.component.html
