+++ src/app/project/ontology-classes/ontology-class-item/ontology-class-item.component.spec.ts
