+++ src/app/project/settings/settings.component.scss
